Update package path, add repo manifest, fix minor bugs
