+++ main.go
@@ -1,20 +1,12 @@
 package main
 
-import (
-	"log"
-
-	api "github.com/crafting-dev/template-go-gin/src"
-)
+import api "github.com/crafting-dev/template-go-gin/pkg/api"
 
 func main() {
 	var ctx api.Context
 
 	ctx.Mode = "release"
 	ctx.Port = "8080"
-
-	if ctx.Port == "" {
-		log.Fatal("PORT must be set")
-	}
 
 	api.Init(ctx)
 }
